split up notification model tests
